requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-ads==28.0.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
proto-plus==1.26.1
grpcio==1.74.0
protobuf==4.25.3
